requirements.txt
requests
pillow
openai
opencv-python
volcengine-python-sdk[ark]
supervision==0.27.0
ui-tars
torchcodec
torch